+++ PHP-Classes/GenHead.class.php
@@ -24,8 +24,18 @@ class GenHead {
         return $out;
     }
 
-    public static function GenIconLinks() {
+    public static function GenLinks(array $arr) {
+        $out = "";
+        foreach ($arr as $i) {
+            $out .= "<link" . ($i['Path'] && $i['Rel'] ? '' : ' DISABLED') . " rel='" . $i['Rel'] . "' href='" . $i['Path'] . "'" . ($i['Type'] ? " type='" . $i['Type'] . "'") . ($i['Size'] ? " size='" . $i['Size'] . "'") . " >";
+        }
+        return $out;
+    }
 
+    public static function GenExtraMetadata(array $arr) {
+        $out = "";
+
+        return $out;
     }
 
     public static function GenScripts(array $arr) {