+++ PHP-Classes/GlobalVars.class.php
@@ -26,6 +26,31 @@ class GlobalVars {
         'Creator' => '@Darnel_Kumar'
     );
 
+    public static $Links = array(
+        array("Path" => "/favicon.ico", "Rel" => "shortcut icon", "Type" => "image/x-icon", "Size" => false),
+        array("Path" => "/Icons/apple-icon-57x57.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "57x57"),
+        array("Path" => "/Icons/apple-icon-60x60.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "60x60"),
+        array("Path" => "/Icons/apple-icon-72x72.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "72x72"),
+        array("Path" => "/Icons/apple-icon-76x76.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "76x76"),
+        array("Path" => "/Icons/apple-icon-114x114.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "114x114"),
+        array("Path" => "/Icons/apple-icon-120x120.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "120x120"),
+        array("Path" => "/Icons/apple-icon-144x144.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "144x144"),
+        array("Path" => "/Icons/apple-icon-152x152.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "152x152"),
+        array("Path" => "/Icons/apple-icon-180x180.png", "Rel" => "apple-touch-icon", "Type" => false, "Size" => "180x180"),
+        array("Path" => "/Icons/android-icon-192x192.png", "Rel" => "icon", "Type" => "image/png", "Size" => "192x192"),
+        array("Path" => "/Icons/favicon-32x32.png", "Rel" => "icon", "Type" => "image/png", "Size" => "32x32"),
+        array("Path" => "/Icons/favicon-96x96.png", "Rel" => "icon", "Type" => "image/png", "Size" => "96x96"),
+        array("Path" => "/Icons/favicon-16x16.png", "Rel" => "icon", "Type" => "image/png", "Size" => "16x16"),
+        array("Path" => "/manifest.json", "Rel" => "manifest", "Type" => false, "Size" => false)
+    );
+
+    public static $ExtraMetadata = array(
+        array("Name" => "viewport", "Content" => "width=device-width, initial-scale=1.0"),
+        array("Name" => "msapplication-TileColor", "Content" => "#3F3F3F"),
+        array("Name" => "msapplication-TileImage", "Content" => "/Icons/ms-icon-144x144.png"),
+        array("Name" => "theme-color", "Content" => "#3F3F3F")
+    );
+
     public static $LogoPath = '/Images/Logo - White Text.svg';
 }
 ?>