+++ index.php
@@ -18,9 +18,8 @@
             "Locale" => "en_GB"
         );
         echo GenHead::GenSocialCards(GlobalVars::$Twitter, $OpenGraph);
+        echo GenHead::GenLinks(GlobalVars::$Links);
     ?>
-    <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Darnel-K | Home</title>
 </head>
 <body id="HomePage">