+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Versioning numbers defined as MAJOR.MINOR.PATCH
 
+## [0.9.0] - 2018-03-03
+### Added
+- GlobalVars class 'Links' Static Variable
+- GenHead class 'GenLinks' function
+- Variety of icons
+
 ## [0.8.0] - 2018-03-03
 ### Changed
 - OpenGraph support added to GenHead class 'GenSocialCards' function